+++ src/routes/+layout.server.ts
@@ -4,33 +4,30 @@ import prisma from "$lib/server/prisma";
 import { redirect } from "@sveltejs/kit";
 
 export const load = (async ({ locals, url }) => {
-  const session = await locals.auth.validate();
+  const path = url.pathname;
 
-  if (url.pathname === "/profile" && !session) {
-    throw redirect(302, "/login");
-  }
-
-  if (url.pathname === "/login" && session) {
+  if ((path === "/login" || path === "/signup") && (locals.session || locals.user)) {
     throw redirect(302, "/profile");
   }
 
-  if (!session) {
-    return {
-      user: null
-    };
-  }
-
-  const user = await prisma.user.findUnique({
-    where: {
-      id: session.user.userId
-    }
-  });
-
-  if (!user) {
+  if (locals.isProtectedRoute && (!locals.session || !locals.user)) {
     throw redirect(302, "/login");
   }
 
+  if (path === "/signup/password" && locals.session) {
+    if (!locals.user) throw redirect(302, "/login");
+    const userKey = await prisma.key.findFirst({
+      where: {
+        id: "username:" + locals.user.username.toLocaleLowerCase()
+      }
+    });
+
+    if (userKey && userKey.hashed_password) {
+      throw redirect(302, "/profile");
+    }
+  }
+
   return {
-    user
+    [locals.user !== null ? "user" : ""]: locals.user
   };
 }) satisfies LayoutServerLoad;