+++ src/routes/api/oauth/+server.ts
@@ -1,25 +1,25 @@
 import { dev } from "$app/environment";
-import { AD_CLIENT_ID, AD_REDIRECT_URI } from "$env/static/private";
+import { MC_AUTH_CLIENT_ID, MC_AUTH_REDIRECT_URI } from "$env/static/private";
 import { createOAuth2AuthorizationUrlWithPKCE } from "@lucia-auth/oauth";
 import { redirect, type RequestHandler } from "@sveltejs/kit";
 
 export const GET: RequestHandler = async ({ cookies }) => {
   // get url to redirect the user to, with the state
-  const [url, codeVerifier, state] = await createOAuth2AuthorizationUrlWithPKCE("https://login.live.com/oauth20_authorize.srf", {
-    clientId: AD_CLIENT_ID,
-    redirectUri: dev ? "http://localhost:5173/api/oauth/microsoft" : AD_REDIRECT_URI,
-    scope: ["XboxLive.signin"],
+  const [url, codeVerifier, state] = await createOAuth2AuthorizationUrlWithPKCE("https://mc-auth.com/oAuth2/authorize", {
+    clientId: MC_AUTH_CLIENT_ID,
+    redirectUri: dev ? "http://localhost:5173/api/oauth/minecraft" : MC_AUTH_REDIRECT_URI,
+    scope: ["profile"],
     codeChallengeMethod: "S256"
   });
 
   // the state can be stored in cookies or localstorage for request validation on callback
-  cookies.set("microsoft_code_verifier", codeVerifier, {
+  cookies.set("minecraft_code_verifier", codeVerifier, {
     path: "/",
     httpOnly: true,
     maxAge: 60 * 60
   });
 
-  cookies.set("microsoft_oauth_state", state, {
+  cookies.set("minecraft_oauth_state", state, {
     path: "/",
     httpOnly: true,
     maxAge: 60 * 60