+++ src/routes/api/oauth/minecraft/+page.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+  import { Loader2 } from "lucide-svelte";
+</script>
+
+<div class="flex h-screen w-screen items-center justify-center">
+  <Loader2 class="h10 w-10 animate-spin" />
+  Processing
+</div>