+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased / Beta]
+
+### Changed
+
+- Changed the OAuth2 provider from Microsoft to Mc-Auth.com.
+- Changed the Login page design.
+
+### Added
+
+- Added Signup page.
+
 ## [1.1.0] - 2023-12-04
 
 ### Changed